Fix CKEditor Text editing (#32)
Prior to this change, the CKEditor IFrame was clicked before switching into it. With the recent changes of CKEditor toolbar, the CKEditor IFrame has become unclickable. Thus, before switching into its body to enter a new text, this change will wait until it's visible which is sufficient to enter text on CKEditor.