Fix US that is dependent from previously executed tests (#33)
Prior tot his change, the US on Tasks attempts to connect with created random user who is admitted as administrator inside the random space which is a shared and reusable space through scenarios. This change will create a new space specifically for it where the first random user is supposed to be an adminstrator of the random space.