fix: Delete Unstable Test Case of deprecated TimeZone Setting UI - MEED-779 (#57)
The TimeZone Drawer settings Test isn't stable due to labels that can
change switch used version of JRE/JDK on server and which maintenance
version (TimeZones localization updates).

In addition this UI is deprecated and should be removed from product
because it's useless. Thus, this Test case has to be removed instead of
investing Time to stabilize it.