Fix: Add adaptations for CSS changes on App Center Popup - MEED-1891 - Meeds-io/MIPs#52 (#114)
This change will adapt CSS Classes change for App Center Popup which has
been changed to re-use a common component for confirmation dialog
instead of a specific one. In addition, the CKEditor loading waiting has
been added in some places to make sure that it's loaded before
proceeding to add content into it.