Update meeds-qa to make it reusable by other test modules
Before this commit, meeds-qa module allow to launch automatic tests for meeds purpose.
To make is reusable by another module, we need to generate a test-jar containing automatic tests to execute.
This commit make this change, so that the test-jar can be imported a module which would use and increase the automatic test coverage