fix: Fix Space Popover Test Case - MEED-2667 - Meeds-io/meeds#1161 (#192)
Prior to this change, the Space Popover Test was failing due to lack of
a step in Scenario. In fact the Popover is closed when the user clicks
on an action. The existing scenario wasn't reopening the popover but was
assuming that it remains open after the user click. This change is
considering this step by reopening the popover after each user click to
test on buttons status.