fix: Enhance Space Management permission configuration - MEED-3300 (#234)
Prior to this change, the automatic tests relies on default
configuration to make sure that simple users can create a space while it
can change on production with an existing database. This change will
make sure to configure the Space Creation Permissions to allow simple
users to create spaces to fit Automatic Tests prerequisite conditions.