fix: Delete Outdated Tests and update XPaths - MEED-7599 - Meeds-io/MIPs#158 (#290)
This change will delete Test case due to the fact that banner isn't
displayed in space card anymore. In addition, the Test case has been
removed due to the changed default spaces list count and that the spaces
aren't displayed all time in screen. In addition to that, the XPath of
Spaces Cards has been computed switch new UI.